+++ src-tauri/src/prefix_tools.rs
@@ -295,7 +295,7 @@ pub async fn install_powershell(
 
     let client = reqwest::Client
         ::builder()
-        .user_agent("star-control/0.3.0")
+        .user_agent("star-control/0.3.1")
         .connect_timeout(std::time::Duration::from_secs(10))
         .timeout(std::time::Duration::from_secs(30))
         .build()